Added a check to validate the constraint on EditRate of the composition and the EditRate within a MainImageSequence's Resource list.

+++ src/test/java/com/netflix/imflibrary/st2067_2/CompositionPlaylistTest.java
@@ -17,8 +17,8 @@ public class CompositionPlaylistTest
         CompositionPlaylist compositionPlaylist = new CompositionPlaylist(inputFile, null);
         Assert.assertTrue(CompositionPlaylist.isCompositionPlaylist(inputFile));
         Assert.assertTrue(compositionPlaylist.toString().length() > 0);
-        Assert.assertEquals(compositionPlaylist.getEditRate().getNumerator(), 24);
-        Assert.assertEquals(compositionPlaylist.getEditRate().getDenominator(), 1);
+        Assert.assertEquals(compositionPlaylist.getEditRate().getNumerator().longValue(), 24);
+        Assert.assertEquals(compositionPlaylist.getEditRate().getDenominator().longValue(), 1);
         Assert.assertEquals(compositionPlaylist.getUUID(), UUID.fromString("682feecb-7516-4d93-b533-f40d4ce60539"));
 
         UUID uuid = UUID.fromString("586286d2-c45f-4b2f-ad76-58eecd0202b4");
